zlib 1.2.0.5

+++ contrib/minizip/crypt.h
@@ -56,7 +56,7 @@ static void init_keys(const char* passwd,unsigned long* pkeys,const unsigned lon
 #define RAND_HEAD_LEN  12
    /* "last resort" source for second part of crypt seed pattern */
 #  ifndef ZCR_SEED2
-#    define ZCR_SEED2 (unsigned long)3141592654L     /* use PI as default pattern */
+#    define ZCR_SEED2 3141592654UL     /* use PI as default pattern */
 #  endif
 
 static int crypthead(passwd, buf, bufSize, pkeys, pcrc_32_tab, crcForCrypting)
